requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-community>=0.0.38
langgraph>=0.0.60
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
pydantic>=2.7.0
